requirements.txt
Django==1.8.2
django-bootstrap3==5.4.0
django-debug-toolbar==1.3.0
django-extensions==1.5.5
fake-factory==0.5.1
ipython==3.1.0
psycopg2==2.6.1
Werkzeug==0.10.4
djangorestframework==3.1.3

